Module grokking_algos::quicksort

Quicksort

quicksort is a module that introduces not only the simple and elegant sorting algorithm we call 'Quicksort' but also the general classification of algorithmic technique it belongs to, known as 'divide and conquer'.

Context

Building from the foundation of problem solving covered in the recursion module, the mechanics of divide and conquer is a method of breaking down complex problems into smaller, simpler ones.

Put simply - divide and conquer algorithms are recursive by their very nature!

We can describe the application of divide and conquer as follows:

Identify the base case

Divide or simplify the problem until it becomes the base case

Examples

Let's rework a simple sum function to illustrate the principle. We can describe the probelm as such:

A function `sum` takes an array,
storing a reference to a sum total.
For every item in the array,
we increase the count of the sum total,
returning total after iteration is complete.

In this, the iterative case, we can envision a for loop mechanic underpinning the approach. However, in the recursive case we must apply those rules of divide and conquer in order to arrive at the appropriate refactoring.

A function `sum` takes an array,
if that array is empty return a sum total of zero.
If that array is not empty, the sum total is equal
to the first value in the array plus the sum of the rest of the array.

Continued Exercises

4.1

Write out the code for the sum function

A - 4.1

Hello world

4.2

Write a recursive function to count the number of items in a list

A - 4.2

Hello world 2

4.3

Find the maximum number in a list

A - 4.3

Hello world 3

4.4

Remembering binary search - it's a divide and conquer algorithm, too. Can you come up with the base case and recursive case for binary search?

A - 4.4

See: recursive_binary_search

Intro to Quicksort

Like selection_sort before it, Quicksort is common and simple sorting algorithm. Its advantage over the former is in its speed - it is a faster algorithm than selection, having an expected runtime performance of O(n log n). In fitting with the theme of this module, it is also a divide and conquer algorithm.

Quicksort in Detail

Given an array of elements, while there are elements, iterate over the elements.
If there is a value at the moment of iteration, let this the pivot element.
Partition the array of elements, into two sub-arrays
(elements less than the pivot and elements greater than the pivot, respectively).
Then, call Quicksort recursively on each of the two sub-arrays.
Finally, combined the sorted sub-arrays on either side of the pivot element and return.

Perils of the Pivot

We have described Quicksort as an algorithm that can be expected to have the performance characteristics of O(n log n) - but, crucially, this is only in the average case. Selection sort, as we saw previously, had a Big O of O(n * n). As it happens this is also the worst case performance for Quicksort, as well.

Interestingly, Quicksort's neighboring sort algorithm: merge sort is actually consistenly O(n log n). The question then is why not just use merge sort over Quicksort all the time?

The answer lies in the value of the Big O constant values between them. Merge sort has a larger constant value than quicksort, making it slower in practice than Quicksort. Additionally, quicksort is generally faster than merge sort in practice because it will tend to perform in the average case more often than the worst case.

But what consitutes the worst case, anyway? The pivot element in Quicksort bears heavy impact on its performance. Thankfully the worst-case is not as frequently encountered in practice, but let's imagine that we have picked the first element in the list as our pivot element (on a related note, most implementations of the algorithm choose either the first or last). Now, let's also imagine we have a sorted list. Because the Quicksort algorithm eagerly begins its execution without checking if the list is sorted, we will still attempt to sort the sorted list.

When this occurs, we will have a longer call stack compared to a middle-point pivot against that same sorted list:

[1] [2] [3] [4] [5] [6] [7] [8]

// NOTE: We pick the first element as the pivot every time
[ ] <1> [2] [3] [4] [5] [6] [7] [8]
        [ ] <2> [3] [4] [5] [6] [7] [8]
            [ ] <3> [4] [5] [6] [7] [8]
                [ ] <4> [5] [6] [7] [8]
                    [ ] <5> [6] [7] [8]
                        [ ] <6> [7] [8]
                            [ ] [7] [8]

Compare this with a middle-point pivot against that same sorted list:

[1] [2] [3] [4] [5] [6] [7] [8]

[1] [2] [3] <4> [5] [6] [7] [8]
[1] <2> [3]         <6> [7] [8]
                    [ ] <7> [8]

In the first case, we have the worst case runtime and in the second we have the best-case. In other words, the stack size of the first is O(n) and the second is O(log n). For both, the time at each level is O(n). Extrapolating this stack size by level time, we have Big O of O(n * n) and O(n log n), respectively.

Exercises

4.5

Printing the value of each element in an array.

A - 4.5

O(n)

4.6

Doubling the value of each element in an array.

A - 4.6

Also, O(n).

4.7

Doubling the value of just the first element in an array.

A - 4.7

O(1)

4.8

Creating a multiplication table with all the elements in the array. So if your array is [2, 3, 7, 8, 10], you first multiply every element by 2, then multiply every element by 3, then by 7, and so on.

A - 4.8

O(n * n) or O(n2)

Functions

_find_max
A max implementation

count
A count implementation

find_max
A function wrapper for _find_max

quick_sort
A quicksort implementation

sum
A sum implementation